worker/requirements.txt
mysql-connector-python
requests
schedule
